requirements.txt
Django>=1.4
South>=0.7.5
locast-core==1.0
